requirements.txt
fastapi
uvicorn
scikit-learn
numpy
pydantic
google-generativeai